+++ R/fastcpd.R
@@ -403,9 +403,9 @@ fastcpd <- function(  # nolint: cyclomatic complexity
     if (family == "lm" || fastcpd_family == "gaussian") {
       # Estimate the variance for each block and then take the average.
       n <- nrow(fastcpd_data)
-      block_size <- 5
+      block_size <- p + 1
       variance_estimation <- rep(NA, n - block_size)
-      for (i in 1:(n - block_size)) {
+      for (i in seq_len(n - block_size)) {
         block_index <- seq_len(block_size) + i - 1
         block_index_lagged <- seq_len(block_size) + i
 
@@ -418,25 +418,35 @@ fastcpd <- function(  # nolint: cyclomatic complexity
         x_t_x <- crossprod(x_block)
         x_t_x_lagged <- crossprod(x_block_lagged)
 
-        block_slope <- solve(crossprod(x_block), crossprod(x_block, y_block))
-        block_lagged_slope <- solve(
-          crossprod(x_block_lagged), crossprod(x_block_lagged, y_block_lagged)
+        tryCatch(
+          expr = {
+            block_slope <-
+              solve(crossprod(x_block), crossprod(x_block, y_block))
+            block_lagged_slope <- solve(
+              crossprod(x_block_lagged),
+              crossprod(x_block_lagged, y_block_lagged)
+            )
+            x_t_x_inv <- solve(x_t_x)
+            x_t_x_inv_lagged <- solve(x_t_x_lagged)
+            inv_product <- x_t_x_inv %*% x_t_x_inv_lagged
+            cross_term_x <- crossprod(
+              x_block[-1, , drop = FALSE],
+              x_block_lagged[-block_size, , drop = FALSE]
+            )
+            cross_term <- inv_product %*% cross_term_x
+            delta_numerator <- crossprod(block_slope - block_lagged_slope)
+            delta_denominator <-
+              sum(diag(x_t_x_inv + x_t_x_inv_lagged - 2 * cross_term))
+            variance_estimation[i] <- delta_numerator / delta_denominator
+          },
+          error = function(e) {
+            variance_estimation[i] <- NA
+            message("Variance estimation failed for block ", i, ".")
+          }
         )
-        x_t_x_inv <- solve(x_t_x)
-        x_t_x_inv_lagged <- solve(x_t_x_lagged)
-        inv_product <- x_t_x_inv %*% x_t_x_inv_lagged
-        cross_term_x <- crossprod(
-          x_block[-1, , drop = FALSE],
-          x_block_lagged[-block_size, , drop = FALSE]
-        )
-        cross_term <- inv_product %*% cross_term_x
-        delta_numerator <- crossprod(block_slope - block_lagged_slope)
-        delta_denominator <-
-          sum(diag(x_t_x_inv + x_t_x_inv_lagged - 2 * cross_term))
-        variance_estimation[i] <- delta_numerator / delta_denominator
       }
 
-      beta <- beta * mean(variance_estimation)
+      beta <- beta * mean(variance_estimation, na.rm = TRUE)
     }
   }